requirements.txt
numpy
pandas
streamlit
bs4
google-api-python-client
python-dotenv